Fix #1961: Split store.rs (4,826 lines) into submodules
store.rs was the largest file in the codebase. Split into store/mod.rs
+ 5 submodules: collections.rs, crud.rs, search.rs, system.rs,
recovery.rs. Struct definitions, internal helpers, and tests remain
in mod.rs. Methods that are now called across submodule boundaries
were promoted to pub(crate) visibility.

Tests added: none (pure structural refactoring, all 300 existing tests pass)
